// src/modules/facilities/enums/facility-unit-item.enum.ts
export enum FacilityUnitItemGroup {
  VIDEO_PRODUCTION = 'VIDEO_PRODUCTION',

  VISION_MIXING = 'VISION_MIXING',
  ROUTING = 'ROUTING',

  CAMERAS = 'CAMERAS',
  LENSES = 'LENSES',
  CAMERA_SUPPORT = 'CAMERA_SUPPORT',

  PLAYOUT_GRAPHICS = 'PLAYOUT_GRAPHICS',
  RECORDING_REPLAY = 'RECORDING_REPLAY',
  MONITORING = 'MONITORING',

  AUDIO = 'AUDIO',
  INTERCOM = 'INTERCOM',
  NETWORKING = 'NETWORKING',

  SYNC_TIMING = 'SYNC_TIMING',
  RF_WIRELESS = 'RF_WIRELESS',

  POWER = 'POWER',
  CABLING = 'CABLING',

  STUDIO_TOOLS = 'STUDIO_TOOLS',
}